Move MultiService to connectors/services/base
