+++ connectors/services/base.py
@@ -3,6 +3,7 @@
 # or more contributor license agreements. Licensed under the Elastic License 2.0;
 # you may not use this file except in compliance with the Elastic License 2.0.
 #
+import asyncio
 import time
 
 from connectors.logger import logger
@@ -59,3 +60,25 @@ class BaseService:
 
         self.errors[0] = errors
         self.errors[1] = first
+
+
+class MultiService:
+    def __init__(self, services):
+        self._services = services
+
+    async def run(self):
+        tasks = [asyncio.create_task(service.run()) for service in self._services]
+
+        done, pending = await asyncio.wait(tasks, return_when=asyncio.FIRST_COMPLETED)
+
+        for task in pending:
+            task.cancel()
+            await task
+
+    def shutdown(self, sig):
+        logger.info(f"Caught {sig}. Graceful shutdown.")
+
+        for service in self._services:
+            logger.debug(f"Shutting down {service.__class__.__name__}...")
+            service.stop()
+            logger.debug(f"Done shutting down {service.__class__.__name__}...")
